pyproject.toml
[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"


[project]
name = "emseepee"
version = "0.1.0"
description = "A collection of MCP servers including Gmail and other productivity tools"
requires-python = ">=3.12"
dependencies = [
	"fastmcp",
	"uvicorn[standard]>=0.32.0",
	"click",
	"fastapi",
	"google-api-python-client>=2.156.0",
	"google-auth-httplib2>=0.2.0",
	"google-auth-oauthlib>=1.2.1",
	"pyyaml>=6.0.0",
]


[project.optional-dependencies]
test = ["pytest>=7.0.0", "pytest-asyncio>=0.21.0", "httpx>=0.24.0"]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "bandit>=1.7.0",
    "safety>=2.0.0",
    "types-PyYAML",
    "types-requests",
]

[project.scripts]
emseepee = "emseepee.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["emseepee", "config"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Gradual typing
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"]  # Skip assert_used test
